Attempt to point the streetview panorama at the launch site
Panoramas are considered valid if they are within a certain
distance of the lat/long of the launch site, but there is
no guarantee that they chose to point towards the pad. This
calculation determines the heading between the two points,
and sets the panorama's heading to point towards it.